refactor(market): Wave 4 — orchestrator rewrite (606→7 LOC page.tsx)
- MarketContent: thin orchestrator (~170 LOC) with 3 hooks
- PortfolioTab: sub-tab router for team/bestand/angebote/watchlist
- MarktplatzTab: sub-tab router for clubverkauf/transferliste/trending
- TrendingSection: extracted from inline JSX
- MarketHeader: extracted balance + title display
- page.tsx: now 7 LOC wrapper
- tsc 0 errors
